[Refactor] - SmallButtonStyle, PromptExampleButtonStyle을 PressedButtonStyle로 통일
